move auth to utils
